# tests/test_specs/swagger2.json
{
  "swagger": "2.0",
  "info": {
    "title": "Swagger 2.0 API",
    "version": "1.0.0"
  },
  "host": "api.example.com",
  "basePath": "/v1",
  "paths": {
    "/users": {
      "get": {
        "summary": "List all users",
        "operationId": "listUsersV1",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "type": "integer",
            "default": 0
          },
          {
            "name": "limit",
            "in": "query",
            "type": "integer",
            "default": 10
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create user (body in parameters)",
        "operationId": "createUserV1",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserInput"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/users/{id}": {
      "delete": {
        "summary": "Delete user",
        "operationId": "deleteUserV1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "id": { "type": "integer" },
        "name": { "type": "string" },
        "email": { "type": "string" }
      }
    },
    "UserInput": {
      "type": "object",
      "properties": {
        "name": { "type": "string" },
        "age": { "type": "integer" }
      }
    }
  }
}